working on new animations
currently preparing the .HTML file for the new animations to come in the .JS file.

The plan is to grow the images of the games / apps when the user scrolls down / that section is visible.

A parent container has been added around the outside with a min-height of the image(s) full grown size. This is to stop other elements shifting and to act as basically a placeholder.